Safari-609.1.11.1/LayoutTests/accessibility/mac/css-speech-speak-expected.txt

Initial
Normal
Spellout
Digits
Literal
No punctuation
Digits and Literal
Spell and Literal
test
This tests that using the CSS3-speech property 'speak-as' works as from a WebCore level (not a platform level, that is up to the platforms)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'spell-out'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal, digits'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal, literal-punctuation'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal, no-punctuation'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal, digits, literal-punctuation'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'spell-out, literal-punctuation'
PASS accessibilityController.focusedElement.childAtIndex(0).speakAs is 'normal, digits'
PASS successfullyParsed is true

TEST COMPLETE
